requirements.txt
streamlit
instaloader
requests
Pillow
